Update MapManager.java
I actually modified this to include possibilities from another plugin that I made for other kinds of special areas, so that those would also appear on the map.  I reverted some of this (hopefully successfully) to reconcile it to work without the other plugin.  The main part that it would affect is the 'if' statements on lines 286 through 302.  I also modified it to update map only if the player's chunk or facing changes, but that might not be worth the effort (lines 134 through 156).